refactor: controller for api specific routes
Placed Api-specific methods in controller inside of Api subfolder.
