+++ app/Http/Controllers/CityController.php
@@ -2,17 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\DestroyCityRequest;
-use App\Http\Requests\StoreCityRequest;
-use App\Http\Requests\UpdateCityRequest;
-use App\Http\Resources\ShowCityResource;
-use App\Models\Airline;
-use App\Models\City;
 use App\Services\AirlineService;
 use App\Services\CityService;
-use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Http\Response;
 use Illuminate\Pagination\CursorPaginator;
 use Illuminate\View\View;
 
@@ -38,42 +30,6 @@ class CityController extends Controller
         return view('city.index', [
             'cities' => $cities,
             'airlines' => $this->airlineService->get()
-        ]);
-    }
-
-    public function store(StoreCityRequest $request): JsonResponse
-    {
-        City::create($request->validated());
-
-        return response()->json([
-            'message' => 'City created successfully.'
-        ], Response::HTTP_CREATED);
-    }
-
-    public function show(City $city): ShowCityResource
-    {
-        return ShowCityResource::make($city);
-    }
-
-    public function update(UpdateCityRequest $request, City $city): JsonResponse
-    {
-        $city->update($request->validated());
-
-        return response()->json([
-            'message' => 'City updated successfully.'
-        ]);
-    }
-
-    /**
-     * 'DestroyCityRequest' is used for checking if deletion confirmation
-     * is to be required
-     */
-    public function destroy(DestroyCityRequest $request, City $city): JsonResponse
-    {
-        $city->delete();
-
-        return response()->json([
-            'message' => 'City deleted successfully.'
         ]);
     }
 }
